projects/biolyceum/src/requirements/esm2.txt
torch
fair-esm
pandas
matplotlib
